requirements.txt
O365==2.0.26
mssql-django==1.4
python-decouple==3.8
schedule==1.2.1
pytest==7.4.3
pytest-mock==3.12.0
python-dotenv==1.0.0
tzlocal==5.2 